+++ R/primitives.R
@@ -1923,11 +1923,6 @@ nvl_if <- function(pred, true, false) {
 
   current_desc <- .current_descriptor(silent = TRUE)
 
-  debug_mode <- is.null(current_desc)
-  if (debug_mode) {
-    current_desc <- local_descriptor()
-  }
-
   desc_true <- local_descriptor()
   true_graph <- trace_fn(true, list(), desc = desc_true, lit_to_array = TRUE)
   desc_false <- local_descriptor()
@@ -1967,8 +1962,7 @@ nvl_if <- function(pred, true, false) {
     list(pred = pred),
     params = list(true_graph = true_graph, false_graph = false_graph),
     infer_fn = infer_fn,
-    desc = current_desc,
-    debug_mode = debug_mode
+    desc = current_desc
   )
   unflatten(true_graph$out_tree, out)
 }
@@ -2025,10 +2019,6 @@ nvl_while <- function(init, cond, body) {
   }
 
   current_desc <- .current_descriptor(silent = TRUE)
-  debug_mode <- is.null(current_desc)
-  if (debug_mode) {
-    current_desc <- local_descriptor()
-  }
 
   desc_cond <- local_descriptor()
 
@@ -2076,8 +2066,7 @@ nvl_while <- function(init, cond, body) {
     args = lapply(flatten(init), maybe_box_arrayish),
     params = list(cond_graph = cond_graph, body_graph = body_graph),
     infer_fn = infer_fn,
-    desc = current_desc,
-    debug_mode = debug_mode
+    desc = current_desc
   )
 
   unflatten(body_graph$out_tree, out)
@@ -2265,10 +2254,6 @@ nvl_scatter <- function(
   }
 
   current_desc <- .current_descriptor(silent = TRUE)
-  debug_mode <- is.null(current_desc)
-  if (debug_mode) {
-    current_desc <- local_descriptor()
-  }
 
   # Trace the update computation function
   # For scatter, the update computation takes 2 scalar arguments (current, update)
@@ -2350,8 +2335,7 @@ nvl_scatter <- function(
       update_computation_graph = update_computation_graph
     ),
     infer_fn = infer_fn,
-    desc = current_desc,
-    debug_mode = debug_mode
+    desc = current_desc
   )
 
   out[[1L]]